# openapi.yaml
openapi: 3.0.3
info:
  title: Koperasi API
  description: |
    API untuk sistem Koperasi Maju Bersama.

    **Konvensi response:** kebanyakan endpoint mengembalikan envelope `{ success, data?, message? }`.
    Error domain mengembalikan `{ success: false, message }` (kadang `errors` untuk validasi Zod).

    Spesifikasi ini dipelihara manual dan diverifikasi CI terhadap `server/openapi/routeManifest.ts`.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ApiDataEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessEnvelope'
        - type: object
          properties:
            data:
              description: Payload utama endpoint
    ApiMessageEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessEnvelope'
        - type: object
          properties:
            message:
              type: string
            id:
              type: string
              description: ID entitas yang baru dibuat (jika ada)
    ApiErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          description: Detail validasi Zod (opsional)
    PaginatedList:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    LoanInput:
      type: object
      required: [memberId, name, amount, tenor, purpose]
      properties:
        memberId: { type: string }
        name: { type: string }
        amount: { type: number, exclusiveMinimum: 0 }
        tenor: { type: number, exclusiveMinimum: 0 }
        purpose: { type: string }
        status:
          type: string
          enum: [Menunggu, Disetujui, Ditolak, Lunas]
          default: Menunggu
        loanDate:
          type: string
          format: date
          description: Tanggal pinjaman opsional (YYYY-MM-DD) untuk data historis; default hari ini
    LoanStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Menunggu, Disetujui, Ditolak, Lunas]
        approvedDate:
          type: string
          format: date
          description: Tanggal pencairan/persetujuan (YYYY-MM-DD) untuk data historis
        rejectionReason:
          type: string
          description: Alasan penolakan pengajuan pinjaman jika status Ditolak
    LoanDisbursementDateInput:
      type: object
      required: [disbursementDate]
      properties:
        disbursementDate:
          type: string
          format: date
          description: Tanggal pencairan pinjaman (YYYY-MM-DD)
    PaymentInput:
      type: object
      required: [amount, method]
      properties:
        amount: { type: number, exclusiveMinimum: 0 }
        method: { type: string }
        paymentDate:
          type: string
          format: date
          description: Tanggal angsuran opsional (YYYY-MM-DD) untuk data historis; default hari ini
    PaymentUpdateInput:
      type: object
      properties:
        amount: { type: number, exclusiveMinimum: 0 }
        method: { type: string }
        paymentDate:
          type: string
          format: date
          description: Tanggal angsuran (YYYY-MM-DD)
...
  responses:
    BadRequest:
      description: Validasi gagal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
...
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
    DataResponse:
      description: Response sukses dengan data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiDataEnvelope'
    MessageResponse:
      description: Response sukses dengan pesan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiMessageEnvelope'
security:
  - bearerAuth: []
paths:
...
  /loans:
    get:
      tags: [loans]
      summary: Daftar pinjaman (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List pinjaman
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiDataEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedList'
    post:
      tags: [loans]
      summary: Ajukan pinjaman baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanInput'
      responses:
        '201':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /loans/{id}/status:
    put:
      tags: [loans]
      summary: Update status pinjaman (approve/reject)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanStatusInput'
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /loans/{id}:
    delete:
      tags: [loans]
      summary: Hapus pinjaman
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/payments:
    get:
      tags: [loans]
      summary: Daftar pencairan dan angsuran (paginated)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Ledger pembayaran pinjaman
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiDataEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedList'
  /loans/{id}/disbursement-date:
    put:
      tags: [loans]
      summary: Ubah tanggal pencairan pinjaman yang sudah disetujui
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanDisbursementDateInput'
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/schedule:
    get:
      tags: [loans]
      summary: Jadwal angsuran pinjaman
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [loans]
      summary: Ganti jadwal angsuran (manual)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  items:
                    type: object
                    required: [installmentNo, dueDate, principalAmount, interestAmount]
                    properties:
                      installmentNo: { type: integer }
                      dueDate: { type: string, format: date }
                      principalAmount: { type: integer }
                      interestAmount: { type: integer }
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/schedule/regenerate:
    post:
      tags: [loans]
      summary: Generate ulang jadwal angsuran dengan rate baru
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interestRate]
              properties:
                interestRate:
                  type: number
                  minimum: 0
                  maximum: 100
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/payments:
    get:
      tags: [loans]
      summary: Riwayat pembayaran satu pinjaman
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
    post:
      tags: [loans]
      summary: Catat pembayaran cicilan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /loans/{id}/payments/{paymentId}:
    put:
      tags: [loans]
      summary: Ubah nominal/tanggal/metode angsuran
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdateInput'
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [loans]
      summary: Hapus angsuran dan hitung ulang alokasi
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...